the-dayz-server/requirements.txt
appdirs==1.4.3
Django==1.10.6
django-crispy-forms==1.6.0
django-markdown-deux==1.0.5
django-pagedown==0.1.1
django-suit==0.2.24
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
docutils==0.13.1
markdown2==2.3.1
olefile==0.44
packaging==16.8
Pillow==4.0.0
PyJWT==1.4.0
pyparsing==2.2.0
six==1.10.0
